CSV Files with Survey Data in 3 Forms
Easy to Follow version is complete, all survey measures, and easy to understand format and responses. This version is not used, but is a useful guide.
CSV 1 is Simplified Questions, some re-scaled, used only for JN 1
CSV 2 is a polished and clear version of the CSV1, but optimized for JN 2.